# docs/openapi/rust.json
{
  "components": {
    "schemas": {
...
      "PatchBookingRequest": {
        "description": "Request body for patching a booking",
        "properties": {
          "end_time": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "start_time": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "ParkHub",
      "url": "https://github.com/nash87/parkhub"
    },
    "description": "Open-source parking lot management system API. Provides endpoints for user authentication, booking management, parking lot administration, credit system, webhooks, and more.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "ParkHub API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bookings/{id}": {
      "delete": {
        "description": "Cancels an active booking and releases the slot.",
        "operationId": "cancel_booking",
        "parameters": [
          {
            "description": "Booking UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Cancel a booking",
        "tags": [
          "Bookings"
        ]
      },
      "get": {
        "description": "Returns a single booking. Only the owner can access it.",
        "operationId": "get_booking",
        "parameters": [
          {
            "description": "Booking UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking found"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get booking by ID",
        "tags": [
          "Bookings"
        ]
      },
      "patch": {
        "description": "Update notes and/or times on a booking. Only the booking owner or an admin may update.",
        "operationId": "update_booking",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Update a booking",
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{id}/invoice/pdf": {
      "get": {
        "operationId": "get_booking_invoice_pdf",
        "parameters": [
          {
            "description": "Booking ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "PDF invoice"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Booking not found"
          }
        },
        "summary": "Download PDF invoice for a booking",
        "tags": [
          "Invoices"
        ]
      }
    },
...
    "/api/v1/users/{id}": {
      "get": {
        "description": "Retrieves any user's profile. Admin/SuperAdmin only.",
        "operationId": "get_user",
        "parameters": [
          {
            "description": "User UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Get user by ID (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local server",
      "url": "/"
    }
  ],
...
}